pyproject.toml
[tool.poetry]
name = "tel"
version = "0.1.0"
description = ""
authors = ["krv006 <rvkamronbek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
drf-spectacular = "^0.27.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
